requirements.txt
flask
flask-login
werkzeug
opencv-python
mysql-connector-python